Move part of EditField's pub API to Editor
