+++ crates/kas-widgets/src/edit/edit_box.rs
@@ -29,9 +29,9 @@ mod EditBox {
     /// ### Messages
     ///
     /// [`SetValueText`] may be used to replace the entire text and
-    /// [`ReplaceSelectedText`] may be used to replace selected text, where
-    /// [`Self::is_editable`]. This triggers the action handlers
-    /// [`EditGuard::edit`] followed by [`EditGuard::activate`].
+    /// [`ReplaceSelectedText`] may be used to replace selected text when this
+    /// widget is [editable](Editor::is_editable). This triggers the action
+    /// handlers [`EditGuard::edit`] followed by [`EditGuard::activate`].
     ///
     /// [`kas::messages::SetScrollOffset`] may be used to set the scroll offset.
     #[autoimpl(Default, Debug where G: trait)]
@@ -233,18 +233,6 @@ mod EditBox {
             self.vert_bar.set_value(cx, self.scroll.offset().1);
         }
 
-        /// Get text contents
-        #[inline]
-        pub fn as_str(&self) -> &str {
-            self.inner.as_str()
-        }
-
-        /// Get the text contents as a `String`
-        #[inline]
-        pub fn clone_string(&self) -> String {
-            self.inner.clone_string()
-        }
-
         /// Clear text contents and undo history
         #[inline]
         pub fn clear(&mut self, cx: &mut EventState) {
@@ -403,18 +391,6 @@ impl<G: EditGuard> EditBox<G> {
         self
     }
 
-    /// Get whether this `EditField` is editable
-    #[inline]
-    pub fn is_editable(&self) -> bool {
-        self.inner.is_editable()
-    }
-
-    /// Set whether this `EditField` is editable
-    #[inline]
-    pub fn set_editable(&mut self, editable: bool) {
-        self.inner.set_editable(editable);
-    }
-
     /// Set whether this `EditBox` uses multi-line mode
     ///
     /// This setting has two effects: the vertical size allocation is increased
@@ -429,26 +405,12 @@ impl<G: EditGuard> EditBox<G> {
         self
     }
 
-    /// True if the editor uses multi-line mode
-    ///
-    /// See also: [`Self::with_multi_line`]
-    #[inline]
-    pub fn multi_line(&self) -> bool {
-        self.inner.multi_line()
-    }
-
     /// Set the text class used
     #[inline]
     #[must_use]
     pub fn with_class(mut self, class: TextClass) -> Self {
         self.inner = self.inner.with_class(class);
         self
-    }
-
-    /// Get the text class used
-    #[inline]
-    pub fn class(&self) -> TextClass {
-        self.inner.class()
     }
 
     /// Adjust the height allocation
@@ -477,27 +439,5 @@ impl<G: EditGuard> EditBox<G> {
     pub fn with_width_em(mut self, min_em: f32, ideal_em: f32) -> Self {
         self.set_width_em(min_em, ideal_em);
         self
-    }
-
-    /// Get whether the widget has edit focus
-    ///
-    /// This is true when the widget is editable and has keyboard focus.
-    #[inline]
-    pub fn has_edit_focus(&self) -> bool {
-        self.inner.has_edit_focus()
-    }
-
-    /// Get whether the input state is erroneous
-    #[inline]
-    pub fn has_error(&self) -> bool {
-        self.inner.has_error()
-    }
-
-    /// Set the error state
-    ///
-    /// When true, the input field's background is drawn red.
-    /// This state is cleared by [`Self::set_string`].
-    pub fn set_error_state(&mut self, cx: &mut EventState, error_state: bool) {
-        self.inner.set_error_state(cx, error_state);
     }
 }
