+++ crates/kas-widgets/src/edit/mod.rs
@@ -12,7 +12,7 @@ mod guard;
 
 pub use edit_box::EditBox;
 pub use edit_field::EditField;
-use editor::Editor;
+pub use editor::Editor;
 pub use guard::*;
 
 use std::fmt::Debug;
